--- src/main/java/com/whl/hotelService/controller/ChatController.java
@@ -1,4 +1,0 @@
-package com.whl.hotelService.controller;
-
-public class ChatController {
-}
